edit-*: disable autofill of auth RFID input
Set type="password" to avoid displaying the docent's RFID, so it's harder
to shoulder surf a valid authentication value.

Prevent the browser from remembering the docent's RFID using the
autocomplete attribute. At least Chrome ignores autocomplete="off", and
fills the field with any known password for the site, which is annoying,
but autocomplete="new-password" avoids that issue.