add script to remove deleted privileges
When a custom field value is deleted in the web UI, Wild Apricot doesn't
remove this value from any user who currently has in, and then
subsequently doesn't show this value in the UI to allow it to be removed
there either. Thus, we must write a custom script to fix WA's problem:-/

The script is checked in for posterity, but disabled so that it can only
be run when explicitly re-enabled, for safety.